Move NewFakeExecutor into a separate file
